add popular component
