+++ frontend/src/components/popular/popular_index_container.js
@@ -1,6 +1,6 @@
 import { connect } from "react-redux";
 import { requestPopular } from "../../actions/show_actions";
-import Splash from "./splash";
+import PopularIndex from "./popular_index";
 
 const msp = (state) => ({
   popular: state.entities.shows,
@@ -10,4 +10,4 @@ const mdp = (dispatch) => ({
   requestPopular: () => dispatch(requestPopular()),
 });
 
-export default connect(msp, mdp)(Splash);
+export default connect(msp, mdp)(PopularIndex);
